# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAX SPEED RACING SHOP Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserPermissions = { [key: string]: unknown };

export interface User {
  id: number;
  username: string;
  name: string;
  role: string;
  permissions?: UserPermissions;
}

export interface LoginResponse {
  success: boolean;
  user: User;
}

export interface CreateUserRequest {
  username: string;
  password: string;
  name: string;
  role: string;
}

export interface UpdateUserRequest {
  username?: string;
  password?: string;
  name?: string;
  role?: string;
}

export type RolePermissions = { [key: string]: unknown };

export interface Role {
  id: number;
  name: string;
  permissions: RolePermissions;
}

export type CreateRoleRequestPermissions = { [key: string]: unknown };

export interface CreateRoleRequest {
  name: string;
  permissions: CreateRoleRequestPermissions;
}

export interface Biaya {
  id: number;
  tanggal: string;
  keterangan: string;
  nilai: number;
}

export interface DashboardSummary {
  totalPenjualan: number;
  totalModal: number;
  totalBiaya: number;
  laba: number;
  labaShared: number;
  totalTransaksi: number;
  cashTotal: number;
  bankTotal: number;
  onlineShopTotal: number;
  kreditTotal: number;
  kreditBelumCair: number;
  onlineShopBelumCair: number;
  biayaItems: Biaya[];
}

export interface ChartData {
  labels: string[];
  penjualan: number[];
  laba: number[];
}

export type PenjualanPaymentMethod =
  (typeof PenjualanPaymentMethod)[keyof typeof PenjualanPaymentMethod];

export const PenjualanPaymentMethod = {
  cash: "cash",
  bank: "bank",
  online_shop: "online_shop",
  kredit: "kredit",
} as const;

export type PenjualanStatusCair =
  | (typeof PenjualanStatusCair)[keyof typeof PenjualanStatusCair]
  | null;

export const PenjualanStatusCair = {
  pending: "pending",
  cair: "cair",
} as const;

export interface Penjualan {
  id: number;
  tanggal: string;
  nomor: number;
  kodeTransaksi: string;
  kodeBarang: string;
  namaBarang: string;
  brand: string;
  harga: number;
  qty: number;
  total: number;
  paymentMethod: PenjualanPaymentMethod;
  nilaiCash?: number | null;
  namaBank?: string | null;
  nilaiBank?: number | null;
  namaOnlineShop?: string | null;
  nilaiOnlineShop?: number | null;
  namaCustomer?: string | null;
  nilaiKredit?: number | null;
  statusCair?: PenjualanStatusCair;
  tanggalCair?: string | null;
  hargaBeli?: number | null;
  noFaktur?: string | null;
  totalModal?: number | null;
}

export type CreatePenjualanRequestPaymentMethod =
  (typeof CreatePenjualanRequestPaymentMethod)[keyof typeof CreatePenjualanRequestPaymentMethod];

export const CreatePenjualanRequestPaymentMethod = {
  cash: "cash",
  bank: "bank",
  online_shop: "online_shop",
  kredit: "kredit",
} as const;

export interface CreatePenjualanRequest {
  tanggal: string;
  kodeBarang: string;
  harga: number;
  qty: number;
  paymentMethod: CreatePenjualanRequestPaymentMethod;
  nilaiCash?: number | null;
  namaBank?: string | null;
  nilaiBank?: number | null;
  namaOnlineShop?: string | null;
  nilaiOnlineShop?: number | null;
  namaCustomer?: string | null;
  nilaiKredit?: number | null;
  noFaktur?: string | null;
}

export type PencairanItemStatus =
  (typeof PencairanItemStatus)[keyof typeof PencairanItemStatus];

export const PencairanItemStatus = {
  pending: "pending",
  cair: "cair",
} as const;

export interface PencairanItem {
  id: number;
  tanggal: string;
  kodeTransaksi: string;
  noFaktur?: string | null;
  namaBarang: string;
  paymentMethod: string;
  namaOnlineShop?: string | null;
  namaCustomer?: string | null;
  nilai: number;
  status: PencairanItemStatus;
  tanggalCair?: string | null;
}

export interface MarkSettledRequest {
  tanggalCair: string;
  namaBank: string;
  rekeningBank: string;
}

export interface TransaksiBank {
  id: number;
  tanggalCair: string;
  noFaktur?: string | null;
  nilai: number;
  sumber: string;
  namaBank: string;
  rekeningBank: string;
  penjualanId: number;
}

export interface CreateBiayaRequest {
  tanggal: string;
  keterangan: string;
  nilai: number;
}

export interface MasterBarang {
  id: number;
  kodeBarang: string;
  namaBarang: string;
  brand: string;
  supplier: string;
  hargaBeli: number;
  hargaJual: number;
}

export interface CreateMasterBarangRequest {
  kodeBarang: string;
  namaBarang: string;
  brand: string;
  supplier: string;
  hargaBeli: number;
  hargaJual: number;
}

export interface MasterBank {
  id: number;
  namaBank: string;
  nomorRekening: string;
  keterangan?: string | null;
}

export interface CreateMasterBankRequest {
  namaBank: string;
  nomorRekening: string;
  keterangan?: string | null;
}

export interface MasterOnlineShop {
  id: number;
  namaOnlineShop: string;
  keterangan?: string | null;
}

export interface CreateMasterOnlineShopRequest {
  namaOnlineShop: string;
  keterangan?: string | null;
}

export interface Customer {
  id: number;
  namaCustomer: string;
  keterangan?: string | null;
}

export interface CreateCustomerRequest {
  namaCustomer: string;
  keterangan?: string | null;
}

export interface ModalItem {
  id: number;
  tanggal: string;
  kodeTransaksi: string;
  kodeBarang: string;
  namaBarang: string;
  brand: string;
  qty: number;
  harga: number;
  total: number;
  hargaBeli: number;
  totalModal: number;
  paymentMethod: string;
}

export interface ModalResponse {
  items: ModalItem[];
  totalPenjualan: number;
  totalModal: number;
  totalLaba: number;
}

export interface ProfitReport {
  startDate: string;
  endDate: string;
  totalPenjualan: number;
  totalModal: number;
  totalBiaya: number;
  laba: number;
  labaShared: number;
}

export interface PaymentBreakdown {
  cash: number;
  bank: number;
  onlineShop: number;
  kredit: number;
  kreditCair: number;
  onlineShopCair: number;
}

export interface TopProduct {
  kodeBarang: string;
  namaBarang: string;
  brand: string;
  totalQty: number;
  totalPenjualan: number;
}

export type GetDashboardSummaryParams = {
  startDate?: string;
  endDate?: string;
};

export type GetDashboardChartParams = {
  period?: GetDashboardChartPeriod;
};

export type GetDashboardChartPeriod =
  (typeof GetDashboardChartPeriod)[keyof typeof GetDashboardChartPeriod];

export const GetDashboardChartPeriod = {
  daily: "daily",
  monthly: "monthly",
} as const;

export type ListPenjualanParams = {
  startDate?: string;
  endDate?: string;
  paymentMethod?: string;
};

export type ListPencairanParams = {
  startDate?: string;
  endDate?: string;
};

export type ListTransaksiBankParams = {
  startDate?: string;
  endDate?: string;
};

export type ListBiayaParams = {
  startDate?: string;
  endDate?: string;
};

export type ListModalParams = {
  startDate?: string;
  endDate?: string;
};

export type GetLaporanProfitParams = {
  startDate?: string;
  endDate?: string;
};

export type GetPaymentBreakdownParams = {
  startDate?: string;
  endDate?: string;
};

export type GetTopProductsParams = {
  limit?: number;
  startDate?: string;
  endDate?: string;
};