# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAX SPEED RACING SHOP Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/chart:
    get:
      operationId: getDashboardChart
      tags: [dashboard]
      summary: Get sales chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, monthly]
            default: monthly
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
...
  /biaya:
    get:
      operationId: listBiaya
      tags: [biaya]
      summary: List expenses
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Biaya"
    post:
      operationId: createBiaya
      tags: [biaya]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBiayaRequest"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Biaya"
  /biaya/{id}:
    delete:
      operationId: deleteBiaya
      tags: [biaya]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /master-bank:
    get:
      operationId: listMasterBank
      tags: [master-bank]
      summary: List banks
      responses:
        "200":
          description: Banks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MasterBank"
    post:
      operationId: createMasterBank
      tags: [master-bank]
      summary: Create bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMasterBankRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterBank"
  /master-bank/{id}:
    delete:
      operationId: deleteMasterBank
      tags: [master-bank]
      summary: Delete bank
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /customer:
    get:
      operationId: listCustomer
      tags: [customer]
      summary: List customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customer]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customer/{id}:
    delete:
      operationId: deleteCustomer
      tags: [customer]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /laporan/profit:
    get:
      operationId: getLaporanProfit
      tags: [laporan]
      summary: Get profit report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
  /laporan/payment-breakdown:
    get:
      operationId: getPaymentBreakdown
      tags: [laporan]
      summary: Get payment method breakdown
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payment breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentBreakdown"
  /laporan/top-products:
    get:
      operationId: getTopProducts
      tags: [laporan]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required: [success, user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        permissions:
          type: object
          additionalProperties: true
      required: [id, username, name, role]
...
    DashboardSummary:
      type: object
      properties:
        totalPenjualan:
          type: number
        totalModal:
          type: number
        totalBiaya:
          type: number
        laba:
          type: number
        labaShared:
          type: number
        totalTransaksi:
          type: integer
        cashTotal:
          type: number
        bankTotal:
          type: number
        onlineShopTotal:
          type: number
        kreditTotal:
          type: number
        kreditBelumCair:
          type: number
        onlineShopBelumCair:
          type: number
        biayaItems:
          type: array
          items:
            $ref: "#/components/schemas/Biaya"
      required:
        - totalPenjualan
        - totalModal
        - totalBiaya
        - laba
        - labaShared
        - totalTransaksi
        - cashTotal
        - bankTotal
        - onlineShopTotal
        - kreditTotal
        - kreditBelumCair
        - onlineShopBelumCair
        - biayaItems
    ChartData:
      type: object
      properties:
        labels:
          type: array
          items:
            type: string
        penjualan:
          type: array
          items:
            type: number
        laba:
          type: array
          items:
            type: number
      required: [labels, penjualan, laba]
...
    Biaya:
      type: object
      properties:
        id:
          type: integer
        tanggal:
          type: string
        keterangan:
          type: string
        nilai:
          type: number
      required: [id, tanggal, keterangan, nilai]
    CreateBiayaRequest:
      type: object
      properties:
        tanggal:
          type: string
        keterangan:
          type: string
        nilai:
          type: number
      required: [tanggal, keterangan, nilai]
...
    MasterBank:
      type: object
      properties:
        id:
          type: integer
        namaBank:
          type: string
        nomorRekening:
          type: string
        keterangan:
          type: string
          nullable: true
      required: [id, namaBank, nomorRekening]
    CreateMasterBankRequest:
      type: object
      properties:
        namaBank:
          type: string
        nomorRekening:
          type: string
        keterangan:
          type: string
          nullable: true
      required: [namaBank, nomorRekening]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        namaCustomer:
          type: string
        keterangan:
          type: string
          nullable: true
      required: [id, namaCustomer]
    CreateCustomerRequest:
      type: object
      properties:
        namaCustomer:
          type: string
        keterangan:
          type: string
          nullable: true
      required: [namaCustomer]
...
    ProfitReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        totalPenjualan:
          type: number
        totalModal:
          type: number
        totalBiaya:
          type: number
        laba:
          type: number
        labaShared:
          type: number
      required: [startDate, endDate, totalPenjualan, totalModal, totalBiaya, laba, labaShared]
    PaymentBreakdown:
      type: object
      properties:
        cash:
          type: number
        bank:
          type: number
        onlineShop:
          type: number
        kredit:
          type: number
        kreditCair:
          type: number
        onlineShopCair:
          type: number
      required: [cash, bank, onlineShop, kredit, kreditCair, onlineShopCair]
    TopProduct:
      type: object
      properties:
        kodeBarang:
          type: string
        namaBarang:
          type: string
        brand:
          type: string
        totalQty:
          type: integer
        totalPenjualan:
          type: number
      required: [kodeBarang, namaBarang, brand, totalQty, totalPenjualan]